Created routes folder and moved relevant files
